In progress updates to join event catch
Needs to be case tested - but working on updating the parameter description fields.
Added lhs for tsn filters and in progress adding in catch 0 function that calculates 0 catches for the events and species you're intested in.
Added in region to event pull (this should eventually be moved over to the get_ema_event